Rename: 로그아웃, 회원탈퇴 로직 분리
